fix(sim-ffi): fix serialization of u128 and u64 fields
